feat: add rpc estimate_cycles
