+++ util/jsonrpc-types/src/lib.rs
@@ -33,7 +33,7 @@ pub use self::blockchain::{
 pub use self::bytes::JsonBytes;
 pub use self::cell::{CellData, CellInfo, CellWithStatus};
 pub use self::debug::{ExtraLoggerConfig, MainLoggerConfig};
-pub use self::experiment::{DaoWithdrawingCalculationKind, DryRunResult};
+pub use self::experiment::{DaoWithdrawingCalculationKind, EstimateCycles};
 pub use self::fee_rate::FeeRateDef;
 pub use self::fixed_bytes::Byte32;
 pub use self::info::{ChainInfo, DeploymentInfo, DeploymentPos, DeploymentState, DeploymentsInfo};
